requirements.txt
flask
requests
psycopg2-binary
torch
transformers
accelerate